+++ lib/droplet_kit.rb
@@ -24,9 +24,13 @@ module DropletKit
   autoload :Account, 'droplet_kit/models/account'
   autoload :DropletUpgrade, 'droplet_kit/models/droplet_upgrade'
   autoload :FloatingIp, 'droplet_kit/models/floating_ip'
+  autoload :Project, 'droplet_kit/models/project'
+  autoload :ProjectAssignment, 'droplet_kit/models/project_assignment'
+  autoload :Links, 'droplet_kit/models/links'
   autoload :Tag, 'droplet_kit/models/tag'
   autoload :TaggedResources, 'droplet_kit/models/tagged_resources'
   autoload :TaggedDropletsResources, 'droplet_kit/models/tagged_droplets_resources'
+  autoload :TaggedImagesResources, 'droplet_kit/models/tagged_images_resources'
   autoload :Volume, 'droplet_kit/models/volume'
   autoload :LoadBalancer, 'droplet_kit/models/load_balancer'
   autoload :StickySession, 'droplet_kit/models/sticky_session'
@@ -38,6 +42,11 @@ module DropletKit
   autoload :FirewallInboundRule, 'droplet_kit/models/firewall_inbound_rule'
   autoload :FirewallOutboundRule, 'droplet_kit/models/firewall_outbound_rule'
   autoload :FirewallPendingChange, 'droplet_kit/models/firewall_pending_change'
+  autoload :CDN, 'droplet_kit/models/cdn'
+  autoload :KubernetesCluster, 'droplet_kit/models/kubernetes_cluster'
+  autoload :KubernetesNodePool, 'droplet_kit/models/kubernetes_node_pool'
+  autoload :KubernetesNode, 'droplet_kit/models/kubernetes_node'
+  autoload :KubernetesOptions, 'droplet_kit/models/kubernetes_options'
 
   # Resources
   autoload :DropletResource, 'droplet_kit/resources/droplet_resource'
@@ -54,6 +63,7 @@ module DropletKit
   autoload :DropletUpgradeResource, 'droplet_kit/resources/droplet_upgrade_resource'
   autoload :FloatingIpResource, 'droplet_kit/resources/floating_ip_resource'
   autoload :FloatingIpActionResource, 'droplet_kit/resources/floating_ip_action_resource'
+  autoload :ProjectResource, 'droplet_kit/resources/project_resource'
   autoload :TagResource, 'droplet_kit/resources/tag_resource'
   autoload :VolumeResource, 'droplet_kit/resources/volume_resource'
   autoload :VolumeActionResource, 'droplet_kit/resources/volume_action_resource'
@@ -61,6 +71,9 @@ module DropletKit
   autoload :LoadBalancerResource, 'droplet_kit/resources/load_balancer_resource'
   autoload :CertificateResource, 'droplet_kit/resources/certificate_resource'
   autoload :FirewallResource, 'droplet_kit/resources/firewall_resource'
+  autoload :CDNResource, 'droplet_kit/resources/cdn_resource'
+  autoload :KubernetesClusterResource, 'droplet_kit/resources/kubernetes_cluster_resource'
+  autoload :KubernetesOptionsResource, 'droplet_kit/resources/kubernetes_options_resource'
 
   # JSON Maps
   autoload :DropletMapping, 'droplet_kit/mappings/droplet_mapping'
@@ -80,9 +93,13 @@ module DropletKit
   autoload :AccountMapping, 'droplet_kit/mappings/account_mapping'
   autoload :DropletUpgradeMapping, 'droplet_kit/mappings/droplet_upgrade_mapping'
   autoload :FloatingIpMapping, 'droplet_kit/mappings/floating_ip_mapping'
+  autoload :ProjectMapping, 'droplet_kit/mappings/project_mapping'
+  autoload :ProjectAssignmentMapping, 'droplet_kit/mappings/project_assignment_mapping'
+  autoload :LinksMapping, 'droplet_kit/mappings/links_mapping'
   autoload :TagMapping, 'droplet_kit/mappings/tag_mapping'
   autoload :TaggedResourcesMapping, 'droplet_kit/mappings/tagged_resources_mapping'
   autoload :TaggedDropletsResourcesMapping, 'droplet_kit/mappings/tagged_droplets_resources_mapping'
+  autoload :TaggedImagesResourcesMapping, 'droplet_kit/mappings/tagged_images_resources_mapping'
   autoload :VolumeMapping, 'droplet_kit/mappings/volume_mapping'
   autoload :LoadBalancerMapping, 'droplet_kit/mappings/load_balancer_mapping'
   autoload :StickySessionMapping, 'droplet_kit/mappings/sticky_session_mapping'
@@ -94,6 +111,11 @@ module DropletKit
   autoload :FirewallInboundRuleMapping, 'droplet_kit/mappings/firewall_inbound_rule_mapping'
   autoload :FirewallOutboundRuleMapping, 'droplet_kit/mappings/firewall_outbound_rule_mapping'
   autoload :FirewallPendingChangeMapping, 'droplet_kit/mappings/firewall_pending_change_mapping'
+  autoload :CDNMapping, 'droplet_kit/mappings/cdn_mapping'
+  autoload :KubernetesClusterMapping, 'droplet_kit/mappings/kubernetes_cluster_mapping'
+  autoload :KubernetesNodePoolMapping, 'droplet_kit/mappings/kubernetes_node_pool_mapping'
+  autoload :KubernetesNodeMapping, 'droplet_kit/mappings/kubernetes_node_mapping'
+  autoload :KubernetesOptionsMapping, 'droplet_kit/mappings/kubernetes_options_mapping'
 
   # Utils
   autoload :PaginatedResource, 'droplet_kit/paginated_resource'
@@ -104,6 +126,7 @@ module DropletKit
   Error = Class.new(StandardError)
   FailedCreate = Class.new(DropletKit::Error)
   FailedUpdate = Class.new(DropletKit::Error)
+  FailedDelete = Class.new(DropletKit::Error)
 
   class RateLimitReached < DropletKit::Error
     attr_accessor :reset_at