+++ lib/droplet_kit/mappings/certificate_mapping.rb
@@ -9,16 +9,21 @@ module DropletKit
       scoped :read do
         property :id
         property :name
+        property :dns_names
         property :not_after
         property :sha1_fingerprint
         property :created_at
+        property :state
+        property :type
       end
 
       scoped :create do
         property :name
+        property :dns_names
         property :private_key
         property :leaf_certificate
         property :certificate_chain
+        property :type
       end
     end
   end