+++ lib/droplet_kit/models/cdn.rb
@@ -1,0 +1,9 @@
+module DropletKit
+  class CDN < BaseModel
+    attribute :id
+    attribute :ttl
+    attribute :origin
+    attribute :endpoint
+    attribute :created_at
+  end
+end