+++ lib/droplet_kit/models/floating_ip.rb
@@ -6,5 +6,13 @@ module DropletKit
 
     # Used for creates
     attribute :droplet_id
+
+    def identifier
+      ip
+    end
+
+    def self.from_identifier(identifier)
+      new(ip: identifier)
+    end
   end
 end