+++ lib/droplet_kit/models/kubernetes_node.rb
@@ -1,0 +1,7 @@
+module DropletKit
+  class KubernetesNode < BaseModel
+    [:id, :name, :status, :created_at, :updated_at].each do |key|
+      attribute(key)
+    end
+  end
+end