+++ lib/droplet_kit/models/links.rb
@@ -1,0 +1,9 @@
+module DropletKit
+  class Links < BaseModel
+    attribute :myself
+    attribute :first
+    attribute :next
+    attribute :prev
+    attribute :last
+  end
+end