+++ lib/droplet_kit/models/region.rb
@@ -6,4 +6,8 @@ module DropletKit
     attribute :available
     attribute :features
   end
+
+  def self.from_identifier(identifier)
+    new(slug: identifier)
+  end
 end