+++ lib/droplet_kit/models/tagged_droplets_resources.rb
@@ -2,5 +2,6 @@ module DropletKit
   class TaggedDropletsResources < BaseModel
     attribute :count
     attribute :last_tagged
+    attribute :last_tagged_uri
   end
 end