+++ lib/droplet_kit/models/tagged_images_resources.rb
@@ -1,0 +1,6 @@
+module DropletKit
+  class TaggedImagesResources < BaseModel
+    attribute :count
+    attribute :last_tagged_uri
+  end
+end